requirements.txt
# ============================================
# MathSolve AI — Python Dependencies
# ============================================

# --- Backend Core ---
fastapi
uvicorn[standard]
sqlalchemy
python-dotenv
pydantic

# --- OCR ---
google-cloud-vision       # Primary OCR (requires GCP billing)
pytesseract               # Fallback OCR (requires tesseract-ocr system package)
Pillow                    # Image preprocessing for OCR

# --- LLM ---
ollama                    # Python client for local Ollama server

# --- RAG ---
llama-index-core
llama-index-readers-file  # PDF reader
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-llms-ollama
chromadb                  # Vector database
sentence-transformers     # Embedding model

# --- Frontend ---
streamlit
requests
pandas                    # For dashboard charts
